requirements.txt
# Core HTTP client for Bybit REST interactions
httpx>=0.27

# Optional: enable private WebSocket client support
websocket-client>=1.7

# Telegram bot integration
python-telegram-bot>=20.7
